+++ example/lib/main.dart
@@ -47,8 +47,8 @@ init() {
       FieldName.EXCEPTION,
       FieldName.STACKTRACE
     ]
-    ..customOpeningDivider = "|"
-    ..customClosingDivider = "|";
+    ..customOpeningDivider = "{"
+    ..customClosingDivider = "}";
 
   FLog.applyConfigurations(config);
 }
@@ -60,7 +60,7 @@ class HomePage extends StatefulWidget {
 
 class _HomePageState extends State<HomePage> {
   //runtime permission
-  final PermissionGroup _permissionGroup = PermissionGroup.storage;
+  final Permission _permissionGroup = Permission.storage;
 
   @override
   void initState() {
@@ -100,37 +100,11 @@ class _HomePageState extends State<HomePage> {
     return Row(
       children: <Widget>[
         _buildButton("Log Event", () {
-          for (int i = 0; i < 2; i++) {
-            if (i % 2 == 0) {
-              FLog.logThis(
-                className: "HomePage",
-                methodName: "_buildRow1",
-                text: "My log",
-                type: LogLevel.INFO,
-                dataLogType: DataLogType.DEVICE.toString(),
-              );
-            } else {
-              FLog.error(
-                text: "My log",
-                dataLogType: "Zubair",
-                className: "Home",
-                exception: Exception("This is a test"),
-              );
-            }
-            FLog.warning(
-              className: "HomePage",
-              methodName: "_buildRow1",
-              text: "My log",
-              dataLogType: "Umair",
-            );
-            // not logged because this LogLevel is lower then default INFO
-            FLog.trace(
-              className: "HomePage",
-              methodName: "_buildRow1",
-              text: "My log",
-              dataLogType: "Umair",
-            );
-          }
+          logInfo();
+          logException();
+          logError();
+          logWarning();
+          logTrace();
         }),
         Padding(padding: EdgeInsets.symmetric(horizontal: 5.0)),
         _buildButton("Print Logs", () {
@@ -212,10 +186,69 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
-  //permission methods:---------------------------------------------------------
-  Future<void> requestPermission(PermissionGroup permission) async {
-    final List<PermissionGroup> permissions = <PermissionGroup>[permission];
-    final Map<PermissionGroup, PermissionStatus> permissionRequestResult =
-        await PermissionHandler().requestPermissions(permissions);
+  // general methods:-----------------------------------------------------------
+  void logInfo() {
+//    FLog.logThis(
+//      className: "HomePage",
+//      methodName: "_buildRow1",
+//      text: "Log text/descritption goes here",
+//      type: LogLevel.INFO,
+//      dataLogType: DataLogType.DEVICE.toString(),
+//    );
+
+    final LogLevel _newLogLevel = null;
+    FLog.getDefaultConfigurations()..activeLogLevel = _newLogLevel;
+    FLog.info(
+        text:
+            'LogLevel set to: ${FLog.getDefaultConfigurations().activeLogLevel}');
   }
+
+  void logException() {
+    try {
+      var result = 9 ~/ 0;
+      print(result);
+    } on Exception catch (exception) {
+      FLog.error(
+        text: "Exception text/descritption goes here",
+        dataLogType: "Exception (the type could be anything)",
+        className: "Home",
+        exception: exception,
+      );
+    }
+  }
+
+  void logError() {
+    try {
+      var string = "Zubair";
+      var index = string[-1];
+    } on Error catch (error) {
+      FLog.error(
+        text: "Error text/descritption goes here",
+        dataLogType: "Error (the type could be anything)",
+        className: "Home",
+        exception: error,
+      );
+    }
+  }
+
+  void logWarning() {
+    FLog.warning(
+      className: "HomePage",
+      methodName: "_buildRow1",
+      text: "Log text/descritption goes here",
+      dataLogType: "Warning (the type could be anything)",
+    );
+  }
+
+  void logTrace() {
+    FLog.trace(
+      className: "HomePage",
+      methodName: "_buildRow1",
+      text: "Log text/descritption goes here",
+      dataLogType: "Trace (the type could be anything)",
+    );
+  }
+
+  //permission methods:---------------------------------------------------------
+  Future<void> requestPermission(Permission permission) => permission.request();
 }