+++ lib/constants/constants.dart
@@ -12,7 +12,15 @@ class Constants {
   static const EXCEPTION_NULL_KEY =
       'Encryption key is not provided, please provide encryption key in LogsConfig';
 
+  /// Null Logs Level Exception
+  static const EXCEPTION_NULL_LOGS_LEVEL =
+      'ActiveLogLevel is not provided or null, please provide activeLogLevel key in LogsConfig';
+
   /// Print Export Messages
+  static const EXCEPTION_NULL_LOGS_LEVEL =
+      'ActiveLogLevel is not provided or null, please provide activeLogLevel key in LogsConfig';
+
+  // Print Messages
   static const PRINT_EXPORT_MSG =
       '\n--------------------------------------------------------------------------------------------\nExporting Logs: This might take a while depending upon the size\n--------------------------------------------------------------------------------------------';
   /// Print Log Messages