+++ CHANGELOG.md
@@ -60,6 +60,16 @@
 * Added Log Level functionality, skipping lower Log Levels
 * Added deleteAllLogsByFilter
 
+## [1.2.2] - Oct 15, 2019
+
+* Upgraded all the dependencies
+* upgraded example to support AndroidX
+
+## [1.3.0-alpha-01] - Mar 23, 2020
+
+* refactored main.dart
+* changed exception to dynamic in-order to handle both exception & error
+